react router setup for landing page and game canvas page
