+++ frontend/src/Home/Home.js
@@ -1,0 +1,17 @@
+import Title from "../App/App.js";
+import Form from "../Login/Form.js";
+
+function Home() {
+  return (
+    <>
+      <div className="title">
+        <h6>Mind Racers</h6>
+      </div>
+      <div className="form-container">
+        <Form />
+      </div>
+    </>
+  );
+}
+
+export default Home;
